pyproject.toml
[build-system]
  build-backend = "uv_build"
  requires = ["uv_build"]

[dependency-groups]
  dev = ["dycw-utilities[test]", "pyright", "rich"]

[project]
  authors = [{ email = "d.wan@icloud.com", name = "Derek Wan" }]
  dependencies = [
    "click >=8.3.1, <9",
    "dycw-utilities >=0.172.3, <1",
    "typed-settings[attrs,click] >=25.3.0, <26",
  ]
  description = "Library to set up cron jobs"
  name = "dycw-setup-cronjob"
  readme = "README.md"
  requires-python = ">= 3.14"
  version = "0.3.0"

  [project.scripts]
    setup-cronjob = "setup_cronjob.cli:_main"

[tool]
  [tool.uv]
    [tool.uv.build-backend]
      module-name = "setup_cronjob"
      module-root = "src"
